Fixed the infinite running error Removed some old stale files in the resources folder
There is an infinite loop in `session._begin_tx` method where a tx is created before a conn is really established.
As a result, if the server went down before the session is closed, then in `session._close`, the session will see a tx should be rolled back and blocks infinitly on `tx._roll_back.consume` as no record will ever return.
This PR makes sure that in `session._begin_tx`, a tx is created after a conn is established.