Fix bookmark comparison
Bookmarks need to be parsed and compared in order to send the most
up-to-date bookmark to an old-ish Neo4j version. Previous versions do
not support multiple bookmarks in metadata. They only support a
single one.

This commit fixes a problem with bookmark parsing for further
comparison. Bookmark format consists of a fixed prefix and an int
identifier. Previous parser assumed that they are colon-separated.