Don't handle raw socket after it's wrapped in connection (#530)
When the connection to the server is lost during or right after sending HELLO, the driver was attempting to get the address from a closed socket which causes an OSError. There is no need to handle the raw socket after a successful handshake anyway as it will be wrapped in a BOLT connection object that already logs networks errors and takes care of the underlying socket.