Fix endless loop in Result.peek with fetch_size=1 (#587)
Additionally Implement `ResultSingle` and `ResultPeek` in TestKit backend.
`ResultSingle` will be deactivated for now as TestKit expects the driver to
raise an error if there is not exactly 1 records in the result stream.
Currently, the driver only warns if there are more and returns None if there are
none. This (breaking) fix will be introduced in 5.0.