Acquire RT without impersonated user if possible
This is an optimization saving a few bytes over the wire.
When there is an explicit database configured for a session, there is no need
to send the impersonated user along with the `ROUTE` request. The only
difference this would make is that insufficient permissions would be noticed
when fetching the RT instead of when performing any actual action on the
session. Since routing is part of the global driver/pool operations and not
coupled to sessions, doing it this way seems like the logical conclusion.