Fast failing discovery on certain errors (#611)
The driver tries its best to fetch a routing table. It will try all possible
routers while skipping routers on most errors. However, there are a few errors
that are caused by the client. Those errors should be surfaced to the user for
a better UX/DX and should fail fast: there is no reason to try another router
if we expect it tho return the same error.

Those errors are:
 - `Neo.ClientError.Database.DatabaseNotFound`
 - all `Neo.ClientError.Security.*`
   - except `Neo.ClientError.Security.AuthorizationExpired`
 - `Neo.ClientError.Transaction.InvalidBookmark`
 - `Neo.ClientError.Transaction.InvalidBookmarkMixture`

This PR also changes auth errors to be properly hydrated when received as HELLO
response.