Improve performance of packstream packer (#622)
The chunking bit of the protocol implementation could be sped up by not
constantly checking if chunking is necessary but postponing the chunking until
it's needed (i.e., before wrapping up a bolt message or when collecting the
byte stream to send it over the wire). This change includes an extra copy of the
data in memory, but since data is usually written in small amounts per buffer
write and chunking occurs rarely compared to this, the cost of constant checking
outweighs the cost of the extra copy for most practical use-cases.