The 4.4 driver only supports Python 3.6+
This change could be considered a breaking change. But in fact it's a fix:
it'll make an attempt to install neo4j with a wrong Python version faster.

The supported Python versions were clearly stated in the documentation from the
start. Furthermore, since some compatibility layers for Python 3.5 have been
removed in this driver version, we better fail fast to indicate the
misconfiguration.