Don't close stale connections while in use (#632)
The pool will close connections marked as stale when trying to pick them up
from the pool. It should not do so while the connection is in use (i.e.,
already borrowed from the pool). In concurrent systems, this would lead to all
sorts of errors caused by the connection being managed from different threads:
 * the thread trying to acquire a connection will close the stale connection on
   encounter
 * while the thread that borrowed it might still be using it, e.g., to fetch
   results or run a query

In fact, in corner cases it is also possible to hit this issue without
concurrency by having a workload that relies on multiple sessions to perform
work simultaneously.