Add `.is_retriable()` to Neo4jError and DriverError
This should help users to implement custom retry policies together with explicit
transactions.
This PR also improves documentation around errors and changes the driver's
session code to also use the new flag for determining when to retry a tx.
This will make sure that it's a reliable feature + simplify the driver's code.