Fix bookmark manager docs (#793)
Results need to be consumed in order for the BMM to pick up the bookmarks.

Note:
Consumption could also happen through session closure, starting another auto-
commit transaction, or maybe, in future protocol versions, will the server send
bookmarks in the first `SUCCESS` message to an auto-commit `RUN`. But not yet.