Improve Noe4jError representation (#806)
* Improve Noe4jError representation

When the error is not received from the DBMS, but instead originates from
somewhere in the driver, it might not have a code and a message. In that case,
we fall back to the default Exception representation.

Related:
 * https://github.com/neo4j/neo4j-python-driver/issues/796

* Make Neo4jError representation more robust

Even though, all errors that come from the server go through
`Noej4jError.hydrate` which (currently) makes sure that `code` *and* `message`
are set, this might change in the future.