Improve error reporting on routing discovery (#870)
Routing drivers (`neo4j[+s[sc]]://` scheme) retry fetching a routing table on
many different errors that are considered not retryable in the context of
transactions. This is to overall improve the driver's stability when connecting
to clusters.

However, this poses the risk of hiding user-input errors (e.g., selecting a
database name that is invalid or doesn't exist). Hence, the driver blacklists a
handful of selected error codes upon which the discovery process is terminated
prematurely, raising the raw error to the user.

We expand the list to include more errors:
 * `Neo.ClientError.Statement.TypeError`, e.g., when  trying to impersonate an
   integer.
 * `Neo.ClientError.Statement.ArgumentError`, e.g., when trying to impersonate
   without the required permissions.
 * `Neo.ClientError.Request.Invalid`, e.g., when trying to select an integer
   database.

Also adding tests for discovery retries and fast failure.