Update AyncCondition to be based on Python 3.11
Most notably, this changes how the async driver behaves when created while no
event loop is running. Previously, this would lead to hard to predict errors
only occurring once the connection pool ran full.

The creation of the async driver will fail directly when created without a
running async loop. This mirrors the behavior of `asyncio.Condition` that was
changed in the same way from Python 3.9 to 3.11.

While this might be considered a breaking change, having a partially working
driver that crashes once under real load poses the real risk of not catching the
mistake until the application is in production, which is much worse than failing
as early as possible.