Improve sync TestKit backend performance
The sync version of the TestKit backend was over 4 times as slow as its async
counterpart. The reason is that the TCPServer shipped with Python hands out an
unbuffered writer by default.

I don't quite understand why this wasn't a problem earlier. The sync backend
code did not change, neither did the TCPServer implementation in the standard
library. It shall remain a mystery *insert mystical sound effect*.