Reduce scope of pip warnings as errors (#954)
The CI has been failing too often recently because of changes in pip and
dependencies. Those failures were mostly irrelevant warning that we turned into
errors for better visibility. However, so far there was never any action we
needed to take other than muting that warning on or the other way.

Therefore, this commit changes it so that all pip installation calls are run
without treating warning as error.