Invalidate writers per database
This should improve the performance of the driver in multi database use-cases.
The driver now only removes a server as a writer for a single database (before
for all databases) if that server returns an error that notifies the driver that
the server is no longer a writer (`Neo.ClientError.Cluster.NotALeader` or
`Neo.ClientError.General.ForbiddenOnReadOnlyDatabase`).