Optional check for concurrent usage errors
Some driver objects (e.g, Sessions, Transactions, Result streams) are not safe
for concurrent use. By default, it will cause hard to interpret errors or, in
the worst case, wrong behavior.

To aid finding such bugs, the driver now detects if the Python interpreter is
running development mode and enables extra locking around those objects. If they
are used concurrently, an error will be raised. The way this is implemented, it
will only cause a one-time overhead when loading the driver's modules if the
checks are disabled.

Obviously, those checks are somewhat expensive as they entail locks (less so in
the async driver). Therefore, the checks are only happening if either
 * Python is started in development mode (`python -X dev ...`) or
 * The environment variable `PYTHONNEO4JDEBUG` is set (to anything non-empty)
   at the time the driver's modules is loaded.