Fix notification polyfill from incomplete GQL status (#1080)
When connected to a server sending GQL statuses instead of notifications, the
driver polyfills `summary.summary_notifications` from the GQL statuses. This PR
fixes a but where the driver would ignore any status that was lacking any of the
possible fields. However, at least the position is optional and should not cause
the status to be swallowed.