Revert breaking change in summary.notification introduced in #1060
When introducing bolt 5.5 support for GQL statuses in the summary, the driver
gained the ability to polyfill old-style notifications from the statuses.
However, `ResultSummary.notifications` used to be `None` when the server didn't
send any notifications. When using the polyfill, the field would be set to `[]`
instead in the same scenario. This PR fixes this.