Fix coroutine not awaited warning
When a certain deadline passed, the driver will not even attempt certain IO
operations. By not calling the coroutine function to then ignore it under these
circumstances (causing the warning), but instead deferring creation of the
coroutine until it is clear we want to await it, we avoid the warning.

Functionally, this changes nothing, but makes the driver less noise (and
probably a better async citizen 🏅).